requirements.txt
python-dotenv
requests
streamlit
PyPDF2
groq




